+++ tests/crossapp_third/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class CrossappThirdConfig(AppConfig):
+    name = 'tests.crossapp_third'
+    label = 'crossapp_third'